requirements.txt
streamlit
pywebview
pytesseract
sentence-transformers
faiss-cpu
pillow
PyPDF2
pdf2image
requests
